+++ src/features/setting/hooks/mutations/useDeleteAccountMutation.ts
@@ -1,16 +1,23 @@
 import { router } from "expo-router";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { deleteAccountFn } from "@/src/apis/loginApi";
 import { removeAccessToken, removeAiProfileId } from "@/src/libs/mmkv";
 import { removeRefreshToken } from "@/src/libs/secureStorage";
 import showToast from "@/src/libs/showToast";
+import axiosInstance from "@/src/libs/axiosInstance";
+import endpoint from "@/src/constants/endpoint";
 import toastMessage from "@/src/constants/toastMessage";
+import type { DeleteAccountDTO } from "../../types/settingTypes";
 
-const useDeleteAccount = () => {
+export const _deleteAccount = async () => {
+  const { data } = await axiosInstance.delete<DeleteAccountDTO>(endpoint.auth.delete);
+  return data;
+};
+
+export const useDeleteAccountMutation = () => {
   const queryClient = useQueryClient();
 
   return useMutation({
-    mutationFn: deleteAccountFn,
+    mutationFn: _deleteAccount,
     onSuccess: async (data) => {
       showToast(toastMessage.deleteAccount.success, "success");
       console.log(data);
@@ -26,5 +33,3 @@ const useDeleteAccount = () => {
     },
   });
 };
-
-export default useDeleteAccount;
